+++ packages/kiarina-utils-common/src/kiarina/utils/common/_helpers/parse_config_string.py
@@ -1,10 +1,10 @@
 from typing import Any
 
-from .._types.config_str import ConfigStr
+from .._types.config_string import ConfigString
 
 
 def parse_config_string(
-    config_str: ConfigStr,
+    config_str: ConfigString,
     *,
     separator: str = "&",
     key_value_separator: str = "=",